requirements.txt
asgiref==3.10.0
distlib==0.4.0
Django==5.2.8
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.20.0
platformdirs==4.5.0
PyJWT==2.10.1
sqlparse==0.5.3
stevedore==5.6.0
virtualenv==20.35.4
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.1
virtualenvwrapper-win==1.2.7
